Use synced folder for /home/vagrant/newspaper_works.
This means that the actual storage for the directory containing the
working directory for newspaper_works is on the *host* and made
available to the guest/vm via vboxsf mount.  Only significant limitation
of this is lack of support for symbolic or hard links, but that does not
stop typical developer workflow for newspaper_works.